Add files via upload
Child Marriage from UNICEF. Author unknown. https://data.unicef.org/wp-content/uploads/2019/10/XLS_Child-marriage-database_Aug-2021.xlsx. This data set and article were about the percent of girls and boys married before they were 18 years old. It talks about the violation of human rights of being married before your are 18. I choose to focus on just girls and comparing different regions around the world. I think this information is very crucial because getting married as a child especially if you are forced into is very dangerous so I think its important to see just how bad this problem is and where it is the most common. The data had a lot of different countries on it and even split it into regions at the bottom of the data sheet. It was missing data for East Asia and Pacific, Europe and Central Asia, and North America. I wish it had the percents for those regions since they are more highly developed and would be interesting to compare them to the less developed countries where the percents were higher.